feat: Add an edit view for scrum details
Create the edit view and add controls for modifying the title and length of the scrum.
Store changes to the scrum in a Data property. Define the property using the @State wrapper because it is needed to mutate the property from within the view.
SwiftUI ovserves @State properties and automatically redraws the view's body when the property changes. This ensures the UI stays up to date as the user manipulates the onscreen controls.

- Create a new SwiftUI View file `DetailEditView.swift`.
- Add a @State property for `data`. Declare the @State property as private so it can be accessed only within the view that it is defined.
- Add Form and a Section with a "Meeting Info" header. The `Form` container automatically adapts the appearance of controls when it renders on different platforms.
- Inside Section, add a TextField for the meeting title.
	- TextField takes a binding to a String. Can use the `$` syntax to create a binding to `data.title`. The current view manages the state of the `data` property.
- Add a HStack and a Slider for the meeting length. In the slider's label closure, add a Text view for accessibility use.
	- The Text view won't appear on screen, but VoiceOver uses it to identify the purpose of the slider.
- Add a Spacer and Text to display the meeting length.