feat: Add a theme picker
Create a sutom interactive view for users to choose the color theme of a meeting view.
Using the theme view, list all the app's themes in a picker control. Then, store the value of the selected theme in a variable that bind to the parent view. In this way, the theme view communicates a theme selection up to the parent view.

ThemePicker.swift
- Create a new SwiftUI View file `ThemePicker.swift`.
- Create a binding to a theme `selection`.
	- This binding communicates changes to the theme within the theme picker, back to the parent view.
- Pass a constant binding to initialize a theme picker in the preview.
- Add a picker with the title String and pass the `selection` binding.
- Add a ForEach to list all cases of Theme.
- Add a ThemeView and tag it with the theme.
	- Can tag subviews when it is needed to be differentiated among them controls like pickers and lilsts. Tag values can be any hashable type like in an enumeration.

Theme.swift
- Add conformance to `CaseIterable` and `Identifiable`, using the theme name as its `id`.
	- Make `Theme` conform to `CaseIterable` so that can easily access all the enumeration's cases.

DetailEditView.swift
- Add a theme picker and pass it a binding to a theme.